Formulations

Description

ExaPF.AbstractFormulation — Type

AbstractFormulation

Second layer of the package, implementing the interface between the first layer (the topology of the network) and the third layer (implementing the callbacks for the optimization solver).

API Reference

Variables

ExaPF.AbstractVariable — Type

AbstractVariable

Variables corresponding to a particular formulation.

ExaPF.State — Type

State <: AbstractVariable

All variables x depending on the variables Control u through a non-linear equation g(x, u) = 0.

ExaPF.Control — Type

Control <: AbstractVariable

Implement the independent variables used in the reduced-space formulation.

ExaPF.Parameters — Type

Parameters <: AbstractVariable

Constant parameters.

Get default values attached to a given variable:

ExaPF.initial — Function

initial(form::AbstractFormulation, var::AbstractVariable)

Return an initial position for the variable var.

Examples

u₀ = initial(form, Control())
x₀ = initial(form, State())

Powerflow solver

ExaPF.powerflow — Function

powerflow(form::AbstractFormulation,
          jacobian::AD.StateJacobianAD,
          x::VT,
          u::VT,
          p::VT;
          kwargs...) where VT <: AbstractVector

Solve the power flow equations g(x, u) = 0 w.r.t. the state x, using a Newton-Raphson algorithm. The power flow equations are specified in the formulation form. The solution x(u) depends on the control u passed in input.

The algorithm stops when a tolerance tol or a maximum number of irations maxiter are reached.

Arguments

form::AbstractFormulation: formulation of the power flow equation

x::AbstractVector: initial state (the value of x is kept constant)

u::AbstractVector: fixed control (the control is determined by the formulation used)

p::AbstractVector: fixed parameters (also dependent on the formulation used)

Optional arguments

tol::Float64 (default 1e-7): tolerance of the Newton-Raphson algorithm.

maxiter::Int (default 20): maximum number of iterations.

verbose_level::Int (default O, max value: 3): verbose level

Constraints

ExaPF.state_constraints — Function

state_constraints(form::AbstractFormulation, g::VT, x::VT, u::VT, p::VT) where {VT<:AbstractVector}

Evaluate the constraints porting on the state x, as a function of x and u. The result is stored inplace, inside g.

ExaPF.power_constraints — Function

power_constraints(form::AbstractFormulation, g::VT, x::VT, u::VT, p::VT) where {VT<:AbstractVector}

Evaluate the constraints on the power production that are not taken into account in

the box constraints on the control u

the box constraints on the state x (implemented in state_constraints)

The result is stored inplace, inside g.

ExaPF.thermal_limit_constraints — Function

thermal_limit_constraints(form::AbstractFormulation, g::VT, x::VT, u::VT, p::VT) where {VT<:AbstractVector}

Evaluate the thermal limit constraints porting on the lines of the network.

The result is stored inplace, inside the vector g.

Admissible range for variables and constraints:

ExaPF.bounds — Function

bounds(form::AbstractFormulation, var::AbstractVariable)

Return the bounds attached to the variable var.

bounds(form::AbstractFormulation, func::Function)

Return the lower and upper bounds attached to a given constraint functional.

Examples

u_min, u_max = bounds(form, Control())
h_min, h_max = bounds(form, power_constraints)

Costs

ExaPF.cost_production — Function

cost_production(form::AbstractFormulation, x, u, p)::Float64

Get operational cost for given state x and control u.

Attributes

ExaPF.AbstractFormAttribute — Type

AbstractFormAttribute

Attributes attached to an AbstractFormulation.

ExaPF.NumberOfState — Type

Number of states attached to a particular formulation.

ExaPF.NumberOfControl — Type

Number of controls attached to a particular formulation.