Description

ExaPF.AD.AbstractADFramework — Type

AbstractADFramework

Automatic differentiation for the compressed Jacobians of the constraints g(x,u) with respect to the state x and the control u (here called design).

TODO: Use dispatch to unify the code of the state and control Jacobian. This is currently not done because the abstraction of the indexing is not yet resolved.

API Reference

ExaPF.AD.StateJacobianAD — Type

StateJacobianAD

Creates an object for the state Jacobian

J::SMT: Sparse uncompressed Jacobian to be used by linear solver. This is either of type SparseMatrixCSC or CuSparseMatrixCSR.

compressedJ::MT: Dense compressed Jacobian used for updating values through AD either of type Matrix or CuMatrix.

coloring::VI: Row coloring of the Jacobian.

t1sseeds::VP: The seeding vector for AD built based on the coloring.

t1sF::VD: Output array of active (AD) type.

x::VT: Input array of passive type. This includes both state and control.

t1sx::VD: Input array of active type.

map::VI: State and control mapping to array x

varx::SubT: View of map on x

t1svarx::SubD: Active (AD) view of map on x

ExaPF.AD.DesignJacobianAD — Type

DesignJacobianAD

Creates an object for the control Jacobian.

J::SMT: Sparse uncompressed Jacobian to be used by linear solver. This is either of type SparseMatrixCSC or CuSparseMatrixCSR.

compressedJ::MT: Dense compressed Jacobian used for updating values through AD either of type Matrix or CuMatrix.

coloring::VI: Row coloring of the Jacobian.

t1sseeds::VP: The seeding vector for AD built based on the coloring.

t1sF::VD: Output array of active (AD) type.

x::VT: Input array of passive type. This includes both state and control.

t1sx::VD: Input array of active type.

map::VI: State and control mapping to array x

varx::SubT: View of map on x

t1svarx::SubD: Active (AD) view of map on x

ExaPF.AD.myseed_kernel_cpu — Function

myseed_kernel_cpu

Seeding on the CPU, not parallelized.

ExaPF.AD.myseed_kernel_gpu — Function

myseed_kernel_gpu

Seeding on GPU parallelized over the ncolor number of duals

ExaPF.AD.seeding — Function

seeding(t1sseeds::CuVector{ForwardDiff.Partials{N,V}}, varx, t1svarx, nbus) where {N, V}

Calling the GPU seeding kernel

seeding(t1sseeds::Vector{ForwardDiff.Partials{N,V}}, varx, t1svarx, nbus) where {N, V}

Calling the CPU seeding kernel

ExaPF.AD.getpartials_cpu — Function

getpartials_cpu(compressedJ, t1sF)

Extract the partials from the AD dual type on the CPU and put it in the compressed Jacobian

ExaPF.AD.getpartials_gpu — Function

getpartials_gpu(compressedJ, t1sF)

Extract the partials from the AD dual type on the GPU and put it in the compressed Jacobian

ExaPF.AD.getpartials — Function

getpartials(compressedJ::CuArray{T, 2}, t1sF, nbus) where T

Calling the GPU partial extraction kernel

getpartials(compressedJ::Array{T, 2}, t1sF, nbus) where T

Calling the GPU partial extraction kernel

ExaPF.AD._uncompress — Function

_uncompress(J_nzVal, J_rowPtr, J_colVal, compressedJ, coloring, nmap)

Uncompress the compressed Jacobian matrix from compressedJ to sparse CSR on the GPU. Only bitarguments are allowed for the kernel. (for GPU only) TODO: should convert to @kernel

ExaPF.AD.uncompress! — Function

uncompress!(J::SparseArrays.SparseMatrixCSC, compressedJ, coloring)

Uncompress the compressed Jacobian matrix from compressedJ to sparse CSC on the CPU.

uncompress!(J::CUDA.CUSPARSE.CuSparseMatrixCSR, compressedJ, coloring)

Uncompress the compressed Jacobian matrix from compressedJ to sparse CSC on the GPU by calling the kernel _uncompress.

ExaPF.AD.residualJacobianAD! — Function

residualJacobianAD!(arrays::StateJacobianAD,
                    residualFunction_polar!,
                    v_m, v_a, ybus_re, ybus_im, pinj, qinj, pv, pq, ref, nbus,
                    timer = nothing)

Update the sparse Jacobian entries using AD. No allocations are taking place in this function.

arrays::StateJacobianAD: Factory created Jacobian object to update

residualFunction_polar: Primal function

v_m, v_a, ybus_re, ybus_im, pinj, qinj, pv, pq, ref, nbus: Inputs both active and passive parameters. Active inputs are mapped to x via the preallocated views.

residualJacobianAD!(arrays::DesignJacobianAD,
                    residualFunction_polar!,
                    v_m, v_a, ybus_re, ybus_im, pinj, qinj, pv, pq, ref, nbus,
                    timer = nothing)

Update the sparse Jacobian entries using AD. No allocations are taking place in this function.

arrays::DesignJacobianAD: Factory created Jacobian object to update

residualFunction_polar: Primal function

v_m, v_a, ybus_re, ybus_im, pinj, qinj, pv, pq, ref, nbus: Inputs both active and passive parameters. Active inputs are mapped to x via the preallocated views.