Evaluators

Description

ExaPF.AbstractNLPEvaluator — Type

AbstractNLPEvaluator

AbstractNLPEvaluator implements the bridge between the problem formulation (see AbstractFormulation) and the optimization solver. Once the problem formulation bridged, the evaluator allows to evaluate in a straightfoward fashion the objective and the different constraints, but also the corresponding gradient and Jacobian objects.

When working in the reduced space, we could use the corresponding ReducedSpaceEvaluator:

ExaPF.ReducedSpaceEvaluator — Type

ReducedSpaceEvaluator{T} <: AbstractNLPEvaluator

Evaluator working in the reduced space corresponding to the control variable u. Once a new point u is passed to the evaluator, the user needs to call the method update! to find the corresponding state x(u) satisfying the equilibrium equation g(x(u), u) = 0.

Taking as input a given AbstractFormulation, the reduced evaluator builds the bounds corresponding to the control u and the state x, and initiate an AutoDiffFactory tailored to the problem. The reduced evaluator could be instantiated on the main memory, or on a specific device (currently, only CUDA is supported).

The bridge to MathOptInterface is encoded by the MOIEvaluator structure:

ExaPF.MOIEvaluator — Type

MOIEvaluator <: MOI.AbstractNLPEvaluator

Bridge from a ExaPF.AbstractNLPEvaluator to a MOI.AbstractNLPEvaluator.

Attributes

nlp::AbstractNLPEvaluator: the underlying ExaPF problem.

hash_x::UInt: hash of the last evaluated variable x

has_hess::Bool (default: false): if true, pass a Hessian structure to MOI.

API Reference

Attributes

ExaPF.Variables — Type

Variables <: AbstractNLPAttribute end

Attribute corresponding to the optimization variables attached to a given AbstractNLPEvaluator.

ExaPF.Constraints — Type

Constraints <: AbstractNLPAttribute end

Attribute corresponding to the constraints attached to a given AbstractNLPEvaluator.

ExaPF.n_variables — Function

n_variables(nlp::AbstractNLPEvaluator)

Get the number of variables in the problem.

ExaPF.n_constraints — Function

n_constraints(nlp::AbstractNLPEvaluator)

Get the number of constraints in the problem.

Callbacks

ExaPF.objective — Function

objective(nlp::AbstractNLPEvaluator, u)::Float64

Evaluate the objective at point u.

ExaPF.gradient! — Function

gradient!(nlp::AbstractNLPEvaluator, g, u)

Evaluate the gradient of the objective at point u. Store the result inplace in the vector g, which should have the same dimension as u.

ExaPF.constraint! — Function

constraint!(nlp::AbstractNLPEvaluator, cons, u)

Evaluate the constraints of the problem at point u. Store the result inplace, in the vector cons.

Note

The vector cons should have the same dimension as the result returned by n_constraints(nlp).

ExaPF.jacobian_structure! — Function

jacobian_structure!(nlp::AbstractNLPEvaluator, rows, cols)

Return the sparsity pattern of the Jacobian matrix.

ExaPF.jacobian! — Function

jacobian!(nlp::ReducedSpaceEvaluator, jac, u)

Evaluate the Jacobian of the constraints at position u. Store the result inplace, in the array jac.

ExaPF.jtprod! — Function

jtprod!(nlp::ReducedSpaceEvaluator, jv, u, v)

Evaluate the transpose Jacobian-vector product $J^{T} v$. The vector jv is modified inplace.

Let (n, m) = n_variables(nlp), n_constraints(nlp).

u is a vector with dimension n

v is a vector with dimension m

jv is a vector with dimension n

ExaPF.hessian! — Function

hessian!(nlp::AbstractNLPEvaluator, hess, u)

Evaluate the Hessian of the problem at point u. Store the result inplace, in the vector hess.

Utilities

ExaPF.reset! — Function

reset!(nlp::AbstractNLPEvaluator)

Reset evaluator nlp to default configuration.

ExaPF.primal_infeasibility — Function

primal_infeasibility(nlp::AbstractNLPEvaluator, u)

Return primal infeasibility associated to current model nlp evaluated at control u.

ExaPF.primal_infeasibility! — Function

primal_infeasibility!(nlp::AbstractNLPEvaluator, cons, u)

Return primal infeasibility associated to current model nlp evaluated at control u. Modify vector cons inplace.