Formulations

Description

ExaPF.AbstractFormulation — Type

AbstractFormulation

Second layer of the package, implementing the interface between the first layer (the topology of the network) and the third layer (implementing the callbacks for the optimization solver).

API Reference

Variables

ExaPF.AbstractVariable — Type

AbstractVariable

Variables corresponding to a particular formulation.

ExaPF.State — Type

State <: AbstractVariable

All variables x depending on the variables Control u through a non-linear equation g(x, u) = 0.

ExaPF.Control — Type

Control <: AbstractVariable

Implement the independent variables used in the reduced-space formulation.

ExaPF.Parameters — Type

Parameters <: AbstractVariable

Constant parameters.

ExaPF.PhysicalState — Type

PhysicalState <: AbstractVariable

All physical variables describing the current physical state of the underlying network.

PhysicalState variables are encoded in a AbstractNetworkBuffer, storing all the physical values needed to describe the current state of the network.

Get default values attached to a given variable:

ExaPF.initial — Function

initial(form::AbstractFormulation, var::AbstractVariable)

Return an initial position for the variable var.

Examples

u₀ = initial(form, Control())
x₀ = initial(form, State())

Attributes

ExaPF.AbstractFormAttribute — Type

AbstractFormAttribute

Attributes attached to an AbstractFormulation.

ExaPF.NumberOfState — Type

Number of states attached to a particular formulation.

ExaPF.NumberOfControl — Type

Number of controls attached to a particular formulation.

ExaPF extends Base.get to query the different attributes of a model:

Base.get — Function

get(form::AbstractFormulation, attr::AbstractFormAttribute)

Return value of attribute attr attached to the particular formulation form.

Examples

get(form, NumberOfState())
get(form, NumberOfControl())

Powerflow solver

ExaPF.powerflow — Function

powerflow(form::AbstractFormulation,
          jacobian::AutoDiff.StateJacobian,
          buffer::AbstractNetworkBuffer;
          kwargs...) where VT <: AbstractVector

Solve the power flow equations g(x, u) = 0 w.r.t. the state x, using a Newton-Raphson algorithm. The powerflow equations are specified in the formulation form. The current state x and control u are specified in buffer. The object buffer is modified inplace.

The algorithm stops when a tolerance tol or a maximum number of irations maxiter are reached.

Arguments

form::AbstractFormulation: formulation of the power flow equation

jacobian::AutoDiff.StateJacobian: Jacobian

buffer::AbstractNetworkBuffer: buffer storing current state x and control u

Optional arguments

linear_solver::AbstractLinearSolver (default DirectSolver()): solver to solve the linear systems J x = y arising at each iteration of the Newton-Raphson algorithm.

tol::Float64 (default 1e-7): tolerance of the Newton-Raphson algorithm.

maxiter::Int (default 20): maximum number of iterations.

verbose_level::Int (default O, max value: 3): verbose level.

ExaPF.power_balance! — Function

power_balance!(form::AbstractFormulation, x::VT, u::VT, p::VT) where {VT<:AbstractVector}

Get power balance at buses, depending on the state x and the control u.

Costs

ExaPF.cost_production — Function

cost_production(form::AbstractFormulation, pg::AbstractVector)::Float64

Get operational cost corresponding to the active power generation specified in the vector pg.

Constraints

Current supported constraints are:

ExaPF.state_constraints — Function

state_constraints(form::AbstractFormulation, cons::AbstractVector, buffer::AbstractNetworkBuffer)

Evaluate the constraints porting on the state x, as a function of x and u. The result is stored inplace, inside g.

ExaPF.power_constraints — Function

power_constraints(form::AbstractFormulation, cons::AbstractVector, buffer::AbstractNetworkBuffer) where {VT<:AbstractVector}

Evaluate the constraints on the power production that are not taken into account in

the box constraints on the control u

the box constraints on the state x (implemented in state_constraints)

The result is stored inplace, inside g.

ExaPF.thermal_limit_constraints — Function

thermal_limit_constraints(form::AbstractFormulation, cons::AbstractVector, buffer::AbstractNetworkBuffer) where {VT<:AbstractVector}

Evaluate the thermal limit constraints porting on the lines of the network.

The result is stored inplace, inside the vector g.

These functions allow to query constraints' attributes:

ExaPF.is_constraint — Function

size_constraint(cons_func::Function)::Bool

Return whether the function cons_func is a supported constraint in the powerflow model.

ExaPF.size_constraint — Function

size_constraint(form::AbstractFormulation, cons_func::Function)::Int

Get number of constraints specified by the function cons_func in the formulation form.

ExaPF.bounds — Function

bounds(form::AbstractFormulation, var::AbstractVariable)

Return the bounds attached to the variable var.

bounds(form::AbstractFormulation, func::Function)

Return a tuple of vectors (lb, ub) specifying the admissible range of the constraints specified by the function cons_func.

Examples

u_min, u_max = bounds(form, Control())
h_min, h_max = bounds(form, power_constraints)

Utils

To ease the integration, the following functions have been imported from MATPOWER. Note that these functions work exclusively on the CPU.

ExaPF.power_balance — Function

power_balance(V, Ybus, Sbus, pv, pq)

Assembly residual function for N-R power flow. In complex form, the balance equations writes

$g(V) = V * (Y_{bus} * V)^* - S_{bus}$

Note

Code adapted from MATPOWER.

ExaPF.residual_jacobian — Function

residual_jacobian(V, Ybus, pv, pq)

Compute the Jacobian w.r.t. the state x of the power balance function power_balance.

Note

Code adapted from MATPOWER.

ExaPF.get_power_injection — Function

get_power_injection(fr, v_m, v_a, ybus_re, ybus_im)

Computes the power injection at node fr. In polar form, the power injection at node i satisfies

\[p_{i} = \sum_{j} v_{i} v_{j} (g_{ij} \cos(\theta_i - \theta_j) + b_{ij} \sin(\theta_i - \theta_j))\]

ExaPF.get_react_injection — Function

get_react_injection(fr, v_m, v_a, ybus_re, ybus_im)

Computes the reactive power injection at node fr. In polar form, the power injection at node i satisfies

\[q_{i} = \sum_{j} v_{i} v_{j} (g_{ij} \sin(\theta_i - \theta_j) - b_{ij} \cos(\theta_i - \theta_j))\]