Description

ExaPF.LinearSolvers.AbstractPreconditioner — Type

AbstractPreconditioner

Preconditioners for the iterative solvers mostly focused on GPUs

API Reference

ExaPF.LinearSolvers.BlockJacobiPreconditioner — Type

BlockJacobiPreconditioner

Creates an object for the block-Jacobi preconditioner

npart::Int64: Number of partitions or blocks

nJs::Int64: Size of the blocks. For the GPUs these all have to be of equal size.

partitions::Vector{Vector{Int64}}:npart` partitions stored as lists

cupartitions: partitions transfered to the GPU

Js: Dense blocks of the block-Jacobi

cuJs: Js transfered to the GPU

map: The partitions as a mapping to construct views

cumap: cumap transferred to the GPU`

part: Partitioning as output by Metis

cupart: part transferred to the GPU

P: The sparse precondition matrix whose values are updated at each iteration

ExaPF.LinearSolvers.update — Function

function update(J::CuSparseMatrixCSR, p, to)

Update the preconditioner p from the sparse Jacobian J in CSR format for the GPU

The dense blocks cuJs are filled from the sparse Jacobian J

To a batch inversion of the dense blocks using CUBLAS

Extract the preconditioner matrix p.P from the dense blocks cuJs

function update(J::CuSparseMatrixCSR, p, to)

Update the preconditioner p from the sparse Jacobian J in CSC format for the CPU

Note that this implements the same algorithm as for the GPU and becomes very slow on CPU with growing number of blocks.

ExaPF.LinearSolvers.build_adjmatrix — Function

build_adjmatrix

Build the adjacency matrix of a matrix A corresponding to the undirected graph

ExaPF.LinearSolvers.fillblock_gpu! — Function

fillblock_gpu

Fill the dense blocks of the preconditioner from the sparse CSC matrix arrays

ExaPF.LinearSolvers.fillP_gpu! — Function

fillblock_gpu

Update the values of the preconditioner matrix from the dense Jacobi blocks