Developer reference

NLP Blocks

The Formulation is decomposed into smaller nonlinear programming (NLP) blocks. Blocks are coupled together using a OPFBlocks structure.

ProxAL.OPFBlocks — Type

OPFBlocks(
    opfdata::OPFData,
    rawdata::RawData;
    modelinfo::ModelParams = ModelParams(),
    backend=JuMPBlockModel,
    algparams::AlgParams = AlgParams()
)

Create a structure OPFBlocks to decompose the original OPF problem specified in opfdata timestep by timestep, by dualizing the ramping constraint. One block corresponds to one optimization subproblem (and hence, to a particular timestep), and the attribute blkCount enumerates the total number of subproblems. The subproblems are specified using AbstractBlockModel objects, allowing to define them either with JuMP (if backend=JuMPBlockModel is chosen) or with ExaPF (backend=ExaBlockModel).

Decomposition by contingencies

By default, OPFBlocks decomposes the problem only timestep by timestep (single-period multiple-contingency scheme), leading to a total of T subproblems. However, if the option algparams.decompCtgs is set to true, the original problem is also decomposed contingency by contingency (single-period single-contingency scheme). In this case the total number of subproblems is T * K (with K the total number of contingencies).

Deporting the resolution on the GPU

When the backend is set to ExaBlockModel (and a CUDA GPU is available), the user could chose to deport the resolution of each subproblem directly on the GPU simply by setting algparams.device=CUDADevice. However, note that we could not instantiate more subproblems on the GPU than the number of GPU available.

Internally, each block is represented as follows.

ProxAL.AbstractBlockModel — Type

AbstractBlockModel

Abstract supertype for the definition of block subproblems.

ProxAL.JuMPBlockModel — Type

JuMPBlockModel(
    blk::Int,
    opfdata::OPFData, raw_data::RawData,
    modelinfo::ModelParams, t::Int, k::Int, T::Int,
)

Use the modeler JuMP to define the optimal power flow inside the block model. This function is called inside the constructor of the structure OPFBlocks, used for decomposition purpose.

Arguments

blk::Int: ID of the block represented by this model

opfdata::OPFData: data used to build the optimal power flow problem.

raw_data::RawData: same data, in raw format

modelinfo::ModelParams: parameters related to specification of the optimization model

t::Int: current time-step. Value should be between 1 and T.

k::Int: current contingency

T::Int: final horizon

ProxAL.ExaBlockModel — Type

ExaBlockModel(
    blk::Int,
    opfdata::OPFData, raw_data::RawData,
    modelinfo::ModelParams, t::Int, k::Int, T::Int;
    device::TargetDevice=CPU,
    nr_tol::Float64=1e-10,
)

Use the package ExaPF to define the optimal power flow inside the block model. This function is called inside the constructor of the structure OPFBlocks, used for decomposition purpose.

Arguments

blk::Int: ID of the block represented by this model

opfdata::OPFData: data used to build the optimal power flow problem.

raw_data::RawData: same data, in raw format

modelinfo::ModelParams: parameters related to specification of the optimization model

t::Int: current time-step. Value should be between 1 and T.

k::Int: current contingency

T::Int: final horizon

ProxAL.init! — Function

init!(block::AbstractBlockModel, algparams::AlgParams)

Init the optimization model by populating the model with variables and constraints.

ProxAL.optimize! — Function

optimize!(block::AbstractBlockModel, x0::AbstractArray, algparams::AlgParams)

Solve the optimization problem, starting from an initial variable x0. The optimization solver is specified in field algparams.optimizer.

ProxAL.set_objective! — Function

set_objective!(
    block::AbstractBlockModel,
    algparams::AlgParams,
    primal::PrimalSolution,
    dual::DualSolution
)

Update the objective inside block's optimization subproblem. The new objective updates the coefficients of the penalty terms, to reflect the new primal and dual solutions passed in the arguments.

ProxAL.get_solution — Function

get_solution(block::AbstractBlockModel, output)

Return the solution of the optimization as a named tuple solution, with fields

status::MOI.TerminationStatus: final status returned by the solver

minimum::Float64: optimal objective found

vm::AbstractArray: optimal values of voltage magnitudes

va::AbstractArray: optimal values of voltage angles

pg::AbstractArray: optimal values of active power generations

qg::AbstractArray: optimal values of reactive power generations

ωt::AbstractArray: optimal values of frequency

st::AbstractArray: optimal values of slack variables

ProxAL.add_variables! — Function

add_variables!(block::AbstractBlockModel, algparams::AlgParams)

Add all optimization variables into the decomposed optimization model block.