Model parameters

Description

ProxAL.ModelParams — Type

ModelParams

Specifies the ACOPF model structure.

Parameter	Description	Default value

num_time_periods::Int	number of time periods	1

num_ctgs::Int	number of line contingencies	0

load_scale::Float64	load multiplication factor	1.0

ramp_scale::Float64	multiply this with $p_{g}^{max}$ to get generator ramping $r_g$	1.0

obj_scale::Float64	objective multiplication factor	1.0e-3

allow_obj_gencost::Bool	model generator cost	true

allow_constr_infeas::Bool	allow constraint infeasibility	false

weight_constr_infeas::Float64	quadratic penalty weight for constraint infeasibilities	1.0

weight_freq_ctrl::Float64	quadratic penalty weight for frequency violations	1.0

weight_ctgs::Float64	linear weight of contingency objective function	1.0

weight_quadratic_penalty_time::Float64	see Formulation	1.0

weight_quadratic_penalty_ctgs::Float64	see Formulation	1.0

case_name::String	name of case file	""

savefile::String	name of save file	""

time_link_constr_type::Symbol	∈ [:penalty, :equality, :inequality] see Formulation	:penalty

ctgs_link_constr_type::Symbol	∈ [:frequency_ctrl, :preventive_penalty, :preventive_equality, :corrective_penalty, :corrective_equality, :corrective_penalty], see Formulation	:preventive_equality

ProxAL.RawData — Type

RawData

Specifies the ACOPF instance data.

bus_arr: read from .bus file

branch_arr: read from .branch file

gen_arr: read from .gen file

costgen_arr: read from .gencost file

pd_arr: read from .Pd file

qd_arr: read from .Qd file

ctgs_arr: read from .Ctgs file

API Reference

Function to load the instance data.

ProxAL.opf_loaddata — Function

opf_loaddata(raw::RawData;
             time_horizon_start::Int=1,
             time_horizon_end::Int=0,
             load_scale::Float64=1.0,
             ramp_scale::Float64=0.0,
             lineOff=Line())

Loads the multi-period ACOPF instance data from raw with the time horizon defined to be [time_horizon_start, time_horizon_end]. Note that time_horizon_end = 0 indicates as many as possible (the number of columns in raw.pd_arr).

All loads in all time periods will be multiplied by load_scale. The ramp_scale is the factor which multiplies $p_{g}^{max}$ to get generator ramping $r_g$. These are set in ModelParams. See Model parameters.

lineOff is a transmission line that can be deleted to represent a contingency.