Description

ExaPF.AutoDiff.AbstractJacobian — Type

AbstractJacobian

Automatic differentiation for the compressed Jacobians of the constraints g(x,u) with respect to the state x and the control u (here called design).

TODO: Use dispatch to unify the code of the state and control Jacobian. This is currently not done because the abstraction of the indexing is not yet resolved.

API Reference

ExaPF.AutoDiff.StateJacobian — Type

StateJacobian

Creates an object for the state Jacobian

J::SMT: Sparse uncompressed Jacobian to be used by linear solver. This is either of type SparseMatrixCSC or CuSparseMatrixCSR.

compressedJ::MT: Dense compressed Jacobian used for updating values through AD either of type Matrix or CuMatrix.

coloring::VI: Row coloring of the Jacobian.

t1sseeds::VP: The seeding vector for AD built based on the coloring.

t1sF::VD: Output array of active (AD) type.

x::VT: Input array of passive type. This includes both state and control.

t1sx::VD: Input array of active type.

map::VI: State and control mapping to array x

varx::SubT: View of map on x

t1svarx::SubD: Active (AD) view of map on x

ExaPF.AutoDiff.ControlJacobian — Type

ControlJacobian

Creates an object for the control Jacobian.

J::SMT: Sparse uncompressed Jacobian to be used by linear solver. This is either of type SparseMatrixCSC or CuSparseMatrixCSR.

compressedJ::MT: Dense compressed Jacobian used for updating values through AD either of type Matrix or CuMatrix.

coloring::VI: Row coloring of the Jacobian.

t1sseeds::VP: The seeding vector for AD built based on the coloring.

t1sF::VD: Output array of active (AD) type.

x::VT: Input array of passive type. This includes both state and control.

t1sx::VD: Input array of active type.

map::VI: State and control mapping to array x

varx::SubT: View of map on x

t1svarx::SubD: Active (AD) view of map on x

ExaPF.AutoDiff.seed_kernel_cpu! — Function

seed_kernel_cpu!

Seeding on the CPU, not parallelized.

ExaPF.AutoDiff.seed_kernel_gpu! — Function

seed_kernel_gpu!

Seeding on GPU parallelized over the ncolor number of duals

ExaPF.AutoDiff.seed_kernel! — Function

seed_kernel!

Calling the GPU seeding kernel

seed_kernel!(t1sseeds::Vector{ForwardDiff.Partials{N,V}}, varx, t1svarx, nbus) where {N, V}

Calling the CPU seeding kernel

ExaPF.AutoDiff.getpartials_kernel_cpu! — Function

getpartials_kernel_cpu!(compressedJ, t1sF)

Extract the partials from the AutoDiff dual type on the CPU and put it in the compressed Jacobian

ExaPF.AutoDiff.getpartials_kernel_gpu! — Function

getpartials_kernel_gpu!(compressedJ, t1sF)

Extract the partials from the AutoDiff dual type on the GPU and put it in the compressed Jacobian

ExaPF.AutoDiff.getpartials_kernel! — Function

getpartials_kernel!(compressedJ::CuArray{T, 2}, t1sF, nbus) where T

Calling the GPU partial extraction kernel

getpartials_kernel!(compressedJ::Array{T, 2}, t1sF, nbus) where T

Calling the CPU partial extraction kernel

ExaPF.AutoDiff.uncompress_kernel_gpu! — Function

uncompress_kernel_gpu!(J_nzVal, J_rowPtr, J_colVal, compressedJ, coloring, nmap)

Uncompress the compressed Jacobian matrix from compressedJ to sparse CSR on the GPU. Only bitarguments are allowed for the kernel. (for GPU only) TODO: should convert to @kernel

ExaPF.AutoDiff.uncompress_kernel! — Function

uncompress_kernel!(J::SparseArrays.SparseMatrixCSC, compressedJ, coloring)

Uncompress the compressed Jacobian matrix from compressedJ to sparse CSC on the CPU.

uncompress_kernel!(J::CUDA.CUSPARSE.CuSparseMatrixCSR, compressedJ, coloring)

Uncompress the compressed Jacobian matrix from compressedJ to sparse CSC on the GPU by calling the kernel uncompress_kernel_gpu!.

ExaPF.AutoDiff.residual_jacobian! — Function

residual_jacobian!(arrays::StateJacobian,
                    residual_polar!,
                    v_m, v_a, ybus_re, ybus_im, pinj, qinj, pv, pq, ref, nbus,
                    timer = nothing)

Update the sparse Jacobian entries using AutoDiff. No allocations are taking place in this function.

arrays::StateJacobian: Factory created Jacobian object to update

residual_polar: Primal function

v_m, v_a, ybus_re, ybus_im, pinj, qinj, pv, pq, ref, nbus: Inputs both active and passive parameters. Active inputs are mapped to x via the preallocated views.

residual_jacobian!(arrays::ControlJacobian,
                    residual_polar!,
                    v_m, v_a, ybus_re, ybus_im, pinj, qinj, pv, pq, ref, nbus,
                    timer = nothing)

Update the sparse Jacobian entries using AutoDiff. No allocations are taking place in this function.

arrays::ControlJacobian: Factory created Jacobian object to update

residual_polar: Primal function

v_m, v_a, ybus_re, ybus_im, pinj, qinj, pv, pq, ref, nbus: Inputs both active and passive parameters. Active inputs are mapped to x via the preallocated views.