AutoDiff

Adjoint

ExaPF.AutoDiff.TapeMemory — Type

TapeMemory{F, S, I}

This object is used as a buffer to compute the adjoint of a given function $h(x)$. It stores internally all intermediate values necessary to compute the adjoint, and cache the stack used in the backward pass.

Note

This structure is largely inspired from ChainRulesCore.jl.

Jacobian

ExaPF.AutoDiff.AbstractJacobian — Type

AbstractJacobian

Automatic differentiation for the compressed Jacobian of any nonlinear constraint $h(x)$.

ExaPF.AutoDiff.ConstantJacobian — Type

AutoDiff.ConstantJacobian <: AbstractJacobian

Creates a constant Jacobian object for a linear function $h(x)$. Using a ConstantJacobian object allows to avoid computing the full Jacobian with AutoDiff when it is not necessary.

ExaPF.AutoDiff.Jacobian — Type

AutoDiff.Jacobian <: AbstractJacobian

Creates an object to compute the Jacobian with ForwardDiff.

Attributes

func::Func: base function to differentiate

var::Union{State,Control}: specify whether we are differentiating w.r.t. the state or the control.

J::SMT: Sparse uncompressed Jacobian to be used by linear solver. This is either of type SparseMatrixCSC or CuSparseMatrixCSR.

compressedJ::MT: Dense compressed Jacobian used for updating values through AD either of type Matrix or CuMatrix.

coloring::VI: Row coloring of the Jacobian.

t1sseeds::VP: The seeding vector for AD built based on the coloring.

t1sF::VD: Output array of active (AD) type.

x::VT: Input array of passive type. This includes both state and control.

t1sx::VD: Input array of active type.

map::VI: State and control mapping to array x

varx::SubT: View of map on x

t1svarx::SubD: Active (AD) view of map on x

ExaPF.AutoDiff.jacobian! — Function

jacobian!(form::AbstractFormulation, jac::AutoDiff.AbstractJacobian, x)

Update inplace the Jacobian $J$ stored inside jac at a new point x.

API reference for the Jacobian

ExaPF.AutoDiff.seed! — Function

seed!

Calling the seeding kernel. Seeding is parallelized over the ncolor number of duals.

ExaPF.AutoDiff.getpartials_kernel! — Function

getpartials_kernel!(compressedJ, t1sF)

Calling the partial extraction kernel. Extract the partials from the AutoDiff dual type on the target device and put it in the compressed Jacobian compressedJ.

ExaPF.AutoDiff.uncompress_kernel! — Function

uncompress_kernel!(J, compressedJ, coloring)

Uncompress the compressed Jacobian matrix from compressedJ to sparse CSC (on the CPU) or CSR (on the GPU).

Hessian

ExaPF.AutoDiff.AbstractHessian — Type

AbstractHessian

Automatic differentiation for the adjoint-Hessian-vector product $λ^⊤ H v$ of any nonlinear constraint $h(x)$.

ExaPF.AutoDiff.Hessian — Type

AutoDiff.Hessian

Creates an object for computing Hessian adjoint tangent projections.

func::Func: base function to differentiate.

host_t1sseeds::VHP: Seeding vector for seeding on the host.

t1sseeds::VP: The seeding vector for AD built based on the coloring.

x::VT: Input array of passive type. This includes both state and control.

t1sF::VD: Output array of active (AD) type.

∂t1sF::VD: Adjoint of the output array.

t1sx::VD: Input array of active type.

∂t1sx::VD: Adjoint of the input array.

map::VI: State and control mapping to array x

varx::SubT: View of map on x

t1svarx::SubD: Active (AD) view of map on x

buffer::Buff: cache for computing the adjoint (could be Nothing)

ExaPF.AutoDiff.adj_hessian_prod! — Function

adj_hessian_prod!(form::AbstractFormulation, H::AutoDiff.AbstractHessian, hv, x, λ, v)

Compute the adjoint-Hessian-vector product $λ^⊤ H v$ at a given point x, and store the result inplace in vector hv.