Evaluators

Description

ExaPF.AbstractNLPEvaluator — Type

AbstractNLPEvaluator

AbstractNLPEvaluator implements the bridge between the problem formulation (see AbstractFormulation) and the optimization solver. Once the problem formulation bridged, the evaluator allows to evaluate in a straightfoward fashion the objective and the different constraints, but also the corresponding gradient and Jacobian objects.

When working in the reduced space, we could use the corresponding ReducedSpaceEvaluator:

ExaPF.ReducedSpaceEvaluator — Type

ReducedSpaceEvaluator{T} <: AbstractNLPEvaluator

Evaluator working in the reduced space corresponding to the control variable u. Once a new point u is passed to the evaluator, the user needs to call the method update! to find the corresponding state x(u) satisfying the equilibrium equation g(x(u), u) = 0.

Taking as input a given AbstractFormulation, the reduced evaluator builds the bounds corresponding to the control u and the state x, and initiate an ADFactory tailored to the problem. The reduced evaluator could be instantiated on the main memory, or on a specific device (currently, only CUDA is supported).

API Reference

ExaPF.n_variables — Function

n_variables(nlp::AbstractNLPEvaluator)

Get the number of variables in the problem.

ExaPF.n_constraints — Function

n_constraints(nlp::AbstractNLPEvaluator)

Get the number of constraints in the problem.

ExaPF.objective — Function

objective(nlp::AbstractNLPEvaluator, u)::Float64

Evaluate the objective at point u.

ExaPF.gradient! — Function

gradient!(nlp::AbstractNLPEvaluator, g, u)

Evaluate the gradient of the objective at point u. Store the result inplace in the vector g, which should have the same dimension as u.

ExaPF.constraint! — Function

constraint!(nlp::AbstractNLPEvaluator, cons, u)

Evaluate the constraints of the problem at point u. Store the result inplace, in the vector cons. The vector cons should have the same dimension as the result returned by n_constraints(nlp).

ExaPF.jacobian_structure! — Function

jacobian_structure!(nlp::AbstractNLPEvaluator, rows, cols)

Return the sparsity pattern of the Jacobian matrix.

ExaPF.jacobian! — Function

jacobian!(nlp::ReducedSpaceEvaluator, jac, u)

Evaluate the Jacobian of the problem at point u. Store the result inplace, in the vector jac.

ExaPF.hessian! — Function

hessian!(nlp::AbstractNLPEvaluator, hess, u)

Evaluate the Hessian of the problem at point u. Store the result inplace, in the vector hess.