Formulations

Description

ExaPF.AbstractFormulation — Type

AbstractFormulation

Second layer of the package, implementing the interface between the first layer (the topology of the network) and the third layer (implementing the callbacks for the optimization solver).

ExaPF.PolarForm — Type

PolarForm{T, IT, VT, MT}

Takes as input a PS.PowerNetwork network and implement the polar formulation model associated to this network. The structure PolarForm stores the topology of the network, as well as the complete indexing used in the polar formulation.

A PolarForm structure can be instantiated both on the host CPU() or directly on the device CUDADevice().

Powerflow solver

ExaPF.powerflow — Function

powerflow(form::AbstractFormulation,
          algo::AbstractNonLinearSolver;
          kwargs...)

powerflow(form::AbstractFormulation,
          jacobian::AutoDiff.Jacobian,
          buffer::AbstractNetworkBuffer,
          algo::AbstractNonLinearSolver;
          kwargs...) where VT <: AbstractVector

Solve the power flow equations $g(x, u) = 0$ w.r.t. the state $x$, using the algorithm specified in algo (NewtonRaphson by default). The initial state $x$ is specified inside buffer. The object buffer is modified inplace in the function.

The algorithm stops when a tolerance tol or a maximum number of iterations maxiter is reached (these parameters being specified in the object algo).

Notes

If only the arguments form and algo are specified to the function, then the Jacobian jacobian and the cache buffer are inferred from the object form.

Arguments

form::AbstractFormulation: formulation of the power flow equation

jacobian::AutoDiff.Jacobian: Jacobian

buffer::AbstractNetworkBuffer: buffer storing current state x and control u

algo::AbstractNonLinearSolver: non-linear solver. Currently only NewtonRaphson is being implemented.

Optional arguments

linear_solver::AbstractLinearSolver (default DirectSolver()): solver to solve the linear systems $J x = y$ arising at each iteration of the Newton-Raphson algorithm.

ExaPF.NewtonRaphson — Type

NewtonRaphson <: AbstractNonLinearSolver

Newton-Raphson algorithm. Used to solve the non-linear equation $g(x, u) = 0$, at a fixed control $u$.

Attributes

maxiter::Int (default 20): maximum number of iterations

tol::Float64 (default 1e-8): tolerance of the algorithm

verbose::Int (default NONE): verbosity level

Constraints

Current supported constraints are:

ExaPF.voltage_magnitude_constraints — Function

voltage_magnitude_constraints(form::AbstractFormulation, cons::AbstractVector, buffer::AbstractNetworkBuffer)

Bounds the voltage magnitudes at PQ nodes:

\[v_{pq}^♭ ≤ v_{pq} ≤ v_{pq}^♯ .\]

The result is stored inplace, inside cons.

Note

The constraints on the voltage magnitudes at PV nodes $v_{pv}$ are taken into account when bounding the control $u$.

ExaPF.active_power_constraints — Function

active_power_constraints(form::AbstractFormulation, cons::AbstractVector, buffer::AbstractNetworkBuffer)

Evaluate the constraints on the active power production at the generators that are not already taken into account in the bound constraints.

\[p_g^♭ ≤ p_g ≤ p_g^♯ .\]

The result is stored inplace, inside the vector cons.

ExaPF.reactive_power_constraints — Function

reactive_power_constraints(form::AbstractFormulation, cons::AbstractVector, buffer::AbstractNetworkBuffer)

Evaluate the constraints on the reactive power production at the generators:

\[q_g^♭ ≤ q_g ≤ q_g^♯ .\]

The result is stored inplace, inside the vector cons.

ExaPF.flow_constraints — Function

flow_constraints(form::AbstractFormulation, cons::AbstractVector, buffer::AbstractNetworkBuffer)

Evaluate the thermal limit constraints porting on the lines of the network. The result is stored inplace, inside the vector cons.

ExaPF.power_balance — Function

power_balance(form::AbstractFormulation, cons::AbstractVector, buffer::AbstractNetworkBuffer)

Evaluate the power balance in the network:

\[g(x, u) = 0 ,\]

corresponding to the balance equations

\[\begin{aligned} p_i &= v_i \sum_{j}^{n} v_j (g_{ij}\cos{(\theta_i - \theta_j)} + b_{ij}\sin{(\theta_i - \theta_j})) \,, & ∀ i ∈ \{PV, PQ\} \\ q_i &= v_i \sum_{j}^{n} v_j (g_{ij}\sin{(\theta_i - \theta_j)} - b_{ij}\cos{(\theta_i - \theta_j})) \,. & ∀ i ∈ \{PQ\} \\ \end{aligned}\]

The result is stored inplace, inside the vector cons.

These functions allow to query constraints' attributes:

ExaPF.is_constraint — Function

size_constraint(cons_func::Function)::Bool

Return whether the function cons_func is a supported constraint in the powerflow model.

ExaPF.size_constraint — Function

size_constraint(form::AbstractFormulation, cons_func::Function)::Int

Get number of constraints specified by the function cons_func in the formulation form.

ExaPF.bounds — Function

bounds(form::AbstractFormulation, var::AbstractVariable)

Return the bounds attached to the variable var.

bounds(form::AbstractFormulation, func::Function)

Return a tuple of vectors (lb, ub) specifying the admissible range of the constraints specified by the function cons_func.

Examples

u_min, u_max = bounds(form, Control())
h_min, h_max = bounds(form, reactive_power_constraints)

ExaPF implements special functions to compute the derivatives of each constraints:

ExaPF.adjoint! — Function

adjoint!(form::AbstractFormulation, pbm::AutoDiff.TapeMemory, adj_h, h, buffer)

Return the adjoint w.r.t. the variables of the network (voltage magnitudes and angles, power injection) for the constraint stored inside the AutoDiff.TapeMemory object pbm. The results are stored directly inside the stack stored inside pbm.

ExaPF.jacobian_transpose_product! — Function

jacobian_transpose_product!(form::AbstractFormulation, pbm::AutoDiff.TapeMemory, buffer, v)

Return the two transpose-Jacobian vector product $(Jᵤ^⊤ v, Jₓ^⊤ v)$ w.r.t. the control $u$ and the state $x$. Store the two resulting vectors directly inside the AutoDiff.TapeMemory pbm.

ExaPF.matpower_jacobian — Function

matpower_jacobian(form::AbstractFormulation, X::Union{State,Control}, cons_func::Function, V::Vector{Complex})
matpower_jacobian(form::AbstractFormulation, X::Union{State,Control}, cons_func::Function, buffer::AbstractNetworkBuffer)

For the constraint cons_func, return the expression of the Jacobian $J$ w.r.t. the state or the control (depending on the argument X), as given by MATPOWER.

ExaPF.matpower_hessian — Function

matpower_hessian(form::AbstractFormulation, cons_func::Function, buffer::AbstractNetworkBuffer, λ::AbstractVector)

For constraint cons_func, return the three matrices $(λ^⊤ H_{xx}, λ^⊤ H_{xu},λ^⊤ H_{uu})$ storing the product of the Hessian tensor $H$ with the vector $\lambda$. The expressions of the Hessian matrices are given by MATPOWER.

ExaPF.jacobian_sparsity — Function

jacobian_sparsity(form::AbstractFormulation, cons_func::Function, X::Union{State,Control})

For the constraint cons_func, return the sparsity pattern of the Jacobian $J$ w.r.t. the state or the control (depending on the argument X).

API Reference

Variables

ExaPF.AbstractVariable — Type

AbstractVariable

Variables corresponding to a particular formulation.

ExaPF.State — Type

State <: AbstractVariable

All variables $x$ depending on the variables Control $u$ through a non-linear equation $g(x, u) = 0$.

ExaPF.Control — Type

Control <: AbstractVariable

Independent variables $u$ used in the reduced-space formulation.

ExaPF.PhysicalState — Type

PhysicalState <: AbstractVariable

All physical variables describing the current physical state of the underlying network.

PhysicalState variables are encoded in a AbstractNetworkBuffer, storing all the physical values needed to describe the current state of the network.

Get default values attached to a given variable:

ExaPF.initial — Function

initial(form::AbstractFormulation, var::AbstractVariable)

Return an initial position for the variable var.

Examples

u₀ = initial(form, Control())
x₀ = initial(form, State())

Attributes

ExaPF.AbstractFormAttribute — Type

AbstractFormAttribute

Attributes attached to an AbstractFormulation.

ExaPF.NumberOfState — Type

Number of states attached to a particular formulation.

ExaPF.NumberOfControl — Type

Number of controls attached to a particular formulation.

ExaPF extends Base.get to query the different attributes of a model:

Base.get — Function

get(form::AbstractFormulation, attr::AbstractFormAttribute)

Return value of attribute attr attached to the particular formulation form.

Examples

get(form, NumberOfState())
get(form, NumberOfControl())

get(pf::AbstractPowerSystem, attr::AbstractNetworkAttribute)

Return value of attribute attr in the AbstractPowerSystem object pf.

get(pf::AbstractPowerSystem, attr::AbstractIndexing)

Return indexing corresponding to a subset of the buses.

Examples

npq = get(pf, NumberOfPQBuses())
npv = get(pf, NumberOfPVBuses())
index_pv = get(pf, PVIndexes()())
index_gen = get(pf, GeneratorIndexes()())

The associated setter is implemented with setvalues!:

ExaPF.setvalues! — Function

setvalues!(form::AbstractFormulation, attr::PS.AbstractNetworkAttribute, values)

Update inplace the attribute's values specified by attr.

Examples

setvalues!(form, ActiveLoad(), new_ploads)
setvalues!(form, ReactiveLoad(), new_qloads)

Costs

ExaPF.cost_production — Function

cost_production(form::AbstractFormulation, buffer::AbstractNetworkBuffer)::Float64

Get operational cost.